+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Fixed style target string field overlapping delete button
 - Moved title text field to advanced options
 - Added advanced option to block messages via notification
+- Added option to change the way response messages are sent